+++ server/core/sqlQueryBuilder.js
@@ -1,3 +1,5 @@
+const moment = require('moment');
+
 class SqlQueryBuilder {
     constructor (knex) {
         this.knex = knex;
@@ -516,15 +518,8 @@ class SqlQueryBuilder {
                 throw new Error('unknown pattern');
             }
 
-            const fyear = firstDate.getFullYear();
-            const fmonth = pad(firstDate.getMonth() + 1);
-            const fday = pad(firstDate.getDate());
-            const lyear = lastDate.getFullYear();
-            const lmonth = pad(lastDate.getMonth() + 1);
-            const lday = pad(lastDate.getDate());
-
-            const queryFirstDate = fyear + '/' + fmonth + '/' + fday;
-            const queryLastDate = lyear + '/' + lmonth + '/' + lday;
+            const queryFirstDate = this.formatDate(firstDate, qb);
+            const queryLastDate = this.formatDate(lastDate, qb);
 
             switch (filter.filterType) {
             case 'equal-pattern':
@@ -560,56 +555,63 @@ class SqlQueryBuilder {
                 });
             }
         } else {
-            const date1 = filter.criterion.date1;
-            const date2 = filter.criterion.date2;
+            let formattedDate1, formattedNextDay, formattedDate2;
 
-            const d1 = new Date(date1);
-            d1.setDate(d1.getDate() + 1);
-            const nextDay = d1.getFullYear() + '-' +
-                pad(d1.getMonth() + 1) + '-' +
-                pad(d1.getDate());
+            if (filter.criterion.date1) {
+                const date1 = new Date(filter.criterion.date1);
+                const nextDay = new Date(filter.criterion.date1);
+                nextDay.setDate(nextDay.getDate() + 1);
+
+                formattedDate1 = this.formatDate(date1, qb);
+                formattedNextDay = this.formatDate(nextDay, qb);
+            }
+
+            if (filter.criterion.date2) {
+                const date2 = new Date(filter.criterion.date2);
+                formattedDate2 = this.formatDate(date2, qb);
+            }
 
             switch (filter.filterType) {
             case 'equal':
                 return applyWhereBuilder(builder => {
-                    builder.where(this.getRef(filter), '>=', date1);
-                    builder.where(this.getRef(filter), '<', nextDay);
+                    builder.where(this.getRef(filter), '>=', formattedDate1);
+                    builder.where(this.getRef(filter), '<', formattedNextDay);
                 });
 
             case 'diferentThan':
                 return applyWhereBuilder(builder => {
-                    builder.where(this.getRef(filter), '<', date1);
-                    builder.orWhere(this.getRef(filter), '>=', nextDay);
+                    builder.where(this.getRef(filter), '<', formattedDate1);
+                    builder.orWhere(this.getRef(filter), '>=', formattedNextDay);
                 });
 
             case 'biggerThan':
                 return applyWhereBuilder(builder => {
-                    builder.where(this.getRef(filter), '>', date1);
+                    builder.where(this.getRef(filter), '>', formattedDate1);
                 });
 
             case 'biggerOrEqualThan':
                 return applyWhereBuilder(builder => {
-                    builder.where(this.getRef(filter), '>=', date1);
+                    builder.where(this.getRef(filter), '>=', formattedDate1);
                 });
 
             case 'lessThan':
                 return applyWhereBuilder(builder => {
-                    builder.where(this.getRef(filter), '<', date1);
+                    builder.where(this.getRef(filter), '<', formattedDate1);
                 });
 
             case 'lessOrEqualThan':
                 return applyWhereBuilder(builder => {
-                    builder.where(this.getRef(filter), '<=', date1);
+                    builder.where(this.getRef(filter), '<=', formattedDate1);
                 });
 
             case 'between':
                 return applyWhereBuilder(builder => {
-                    builder.whereBetween(this.getRef(filter), [date1, date2]);
+                    builder.whereBetween(this.getRef(filter), [formattedDate1, formattedDate2]);
                 });
 
             case 'notBetween':
                 return applyWhereBuilder(builder => {
-                    builder.whereNotBetween(this.getRef(filter), [date1, date2]);
+                    builder.whereNotBetween(this.getRef(filter), [formattedDate1, formattedDate2]);
                 });
 
             case 'null':
@@ -625,6 +627,16 @@ class SqlQueryBuilder {
         }
 
         throw new Error('Invalid filter type : ' + filter.filterType);
+    }
+
+    formatDate (date, qb) {
+        let format = 'YYYY-MM-DD';
+
+        if (qb && qb.client && qb.client.config && qb.client.config.client === 'oracledb') {
+            format = 'DD-MMM-YYYY';
+        }
+
+        return moment(date).format(format);
     }
 }